Added dependencies and outline to notebooko

+++ db_schema.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS coffee (
 	id INT PRIMARY KEY,
 	name TEXT,
 	price FLOAT,
-	star_rating INT,
+	star_rating FLOAT,
 	source TEXT,
 	url TEXT
 	);
